Add mass download
Listenes to onDetermeningFilename to get the files name for the
extention.
while suggesting a filename I can't use the \\ path seperator because then
chrome thinks the "file path is too long".
while trying to fix that I trimmed the course number from the course name,
I should probably make that optional later on.