Simplify button building
I was not going to do this at first because this didn't seem it was going
to be readable but it actually removes a lot of the noise that was
surrounding building the buttons and only shows the important parts.

I'd like to be able to import button_html directly but that probably
requires a build system which is really out of scope for what I'm doing
here.

I also deleted an unused function that got away - whoops!