Correctly download redirecting files
Files that open in a new tab open a window with the <a> onclick method.
By extracting the link from the onclick method and using regex matching We
can download the file directly.
Also fixed other regex matching (match[0] is the whole string) and the
reduce. reduce params are (function,_initial value_) and not
(function,starting_index) like I thought